chore: move mocks to new location
